+++ Licenta.Core/Entities/User.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Licenta.Core.Entities;
 
-public class User : IdentityUser<int>
+public class User : IdentityUser<long>
 {
     public string RefreshToken { get; set; }
     public virtual ICollection<UserRole> UserRoles { get; set; }
