+++ Licenta.Core/Entities/UserRole.cs
@@ -2,7 +2,7 @@
 
 namespace Licenta.Core.Entities;
 
-public class UserRole : IdentityUserRole<int>
+public class UserRole : IdentityUserRole<long>
 {
     public virtual User User { get; set; }
     public virtual Role Role { get; set; }
